# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Creators Dashboard dashboard API
  description: |
    A Flask + MongoDB API for the dashboard service.
    
    ## Authentication
    All endpoints (except `/dev-login` and `/metrics`) require JWT authentication via the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    
    For local development, obtain a token from the `/dev-login` endpoint.
  version: 0.1.0
  contact:
    name: Agile Crafts People
    email: info@agilecrafts.com
  license:
    name: Apache-2.0
servers:
  - url: http://localhost:9100
    description: Local development server
...
paths:
...
  /api/Dashboard:
    get:
      tags:
        - Dashboard
      summary: Get all Dashboard documents
      description: Retrieve infinite scroll batch of sorted, filtered Dashboard documents
      operationId: getDashboards
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Dashboard
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, status, created.at_time, saved.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved controls
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Control'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Dashboard
      summary: Create a new Dashboard document
      description: Create a new Dashboard document (requires staff or admin role)
      operationId: createDashboard
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardInput'
      responses:
        '201':
          description: Dashboard created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Dashboard/{DashboardId}:
    get:
      tags:
        - Dashboard
      summary: Get Dashboard by DashboardId
      description: Retrieve a specific Dashboard document by its DashboardId
      operationId: getDashboard
      security:
        - bearerAuth: []
      parameters:
        - name: DashboardId
          in: path
          required: true
          description: The Dashboard ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Dashboard
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags:
        - Dashboard
      summary: Update Dashboard document
      description: Update a Dashboard document (requires admin role). Cannot update _id, created, or saved fields.
      operationId: updateDashboard
      security:
        - bearerAuth: []
      parameters:
        - name: DashboardId
          in: path
          required: true
          description: The Dashboard ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DashboardUpdate'
      responses:
        '200':
          description: Dashboard updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dashboard'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Post:
    get:
      tags:
        - Post
      summary: Get all Post documents
      description: Retrieve infinite scroll batch of sorted, filtered Post documents
      operationId: getPosts
      security:
        - bearerAuth: []
      parameters:
        - name: name
          in: query
          required: false
          description: Optional name filter (supports partial matches, case-insensitive)
          schema:
            type: string
            example: my-Post
        - name: after_id
          in: query
          required: false
          description: Cursor for infinite scroll (ID of last item from previous batch, omit for first request)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
        - name: limit
          in: query
          required: false
          description: Items per page (max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
            example: 10
        - name: sort_by
          in: query
          required: false
          description: Field to sort by (name, description, created.at_time)
          schema:
            type: string
            default: name
            example: name
        - name: order
          in: query
          required: false
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: asc
            example: asc
      responses:
        '200':
          description: Successfully retrieved Posts
          content:
            application/json:
              schema:
                type: object
                required:
                  - items
                  - limit
                  - has_more
                  - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  limit:
                    type: integer
                    example: 10
                  has_more:
                    type: boolean
                    example: true
                  next_cursor:
                    type: string
                    nullable: true
                    pattern: '^[0-9a-fA-F]{24}$'
                    example: 507f1f77bcf86cd799439011
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags:
        - Post
      summary: Create a new Post document
      description: Create a new Post document (requires staff or admin role)
      operationId: createPost
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostInput'
      responses:
        '201':
          description: Post created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    pattern: '^[0-9a-fA-F]{24}$'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/Post/{PostId}:
    get:
      tags:
        - Post
      summary: Get Post by PostId
      description: Retrieve a specific Post document by its PostId
      operationId: getPost
      security:
        - bearerAuth: []
      parameters:
        - name: PostId
          in: path
          required: true
          description: The Post ID (MongoDB ObjectId)
          schema:
            type: string
            pattern: '^[0-9a-fA-F]{24}$'
            example: 507f1f77bcf86cd799439011
      responses:
        '200':
          description: Successfully retrieved Post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from `/dev-login` endpoint.
        Include the token in the Authorization header: `Bearer <token>`
...
  responses:
    BadRequest:
      description: Bad Request - Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
          example: Unauthorized
    Breadcrumb:
      type: object
      description: Audit trail information
      additionalProperties: false
      required:
        - from_ip
        - by_user
        - at_time
        - correlation_id
      properties:
        from_ip:
          type: string
          format: ip
          description: HTTP request remote IP address
          example: 192.168.1.1
        at_time:
          type: string
          format: date-time
          description: Timestamp of the action
          example: "2024-01-15T10:00:00Z"
        by_user:
          type: string
          description: User ID who performed the action
          pattern: '^[^\s]{1,40}$'
          example: user-123
        correlation_id:
          type: string
          description: Logging correlation ID
          pattern: '^[^\s]{1,40}$'
          example: corr-abc123
    Dashboard:
      type: object
      description: Dashboard domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
        - saved
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Dashboard name
          pattern: '^[^\s]{1,40}$'
          example: my-Dashboard
        description:
          type: string
          description: Dashboard description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Dashboard document
        status:
          type: string
          description: Dashboard status
          enum:
            - active
            - archived
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
        saved:
          $ref: '#/components/schemas/Breadcrumb'
    DashboardInput:
      type: object
      description: Input for creating a Dashboard document
      properties:
        name:
          type: string
          description: Dashboard name
          pattern: '^[^\s]{1,40}$'
          example: my-Dashboard
        description:
          type: string
          description: Dashboard description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Dashboard document
        status:
          type: string
          description: Dashboard status
          enum:
            - active
            - archived
          example: active
    DashboardUpdate:
      type: object
      description: Fields to update in a Dashboard document (cannot include _id, created, saved)
      properties:
        name:
          type: string
          description: Dashboard name
          pattern: '^[^\s]{1,40}$'
        description:
          type: string
          description: Dashboard description
          pattern: '^[^\t\n]{0,255}$'
        status:
          type: string
          description: Dashboard status
          enum:
            - active
            - archived
    Post:
      type: object
      description: Post domain document
      additionalProperties: false
      required:
        - _id
        - name
        - created
      properties:
        _id:
          type: string
          description: MongoDB document ID
          pattern: '^[0-9a-fA-F]{24}$'
          example: 507f1f77bcf86cd799439011
        name:
          type: string
          description: Post name
          pattern: '^[^\s]{1,40}$'
          example: my-Post
        description:
          type: string
          description: Post description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Post document
        status:
          type: string
          description: Post status
          example: active
        created:
          $ref: '#/components/schemas/Breadcrumb'
    PostInput:
      type: object
      description: Input for creating a Post document
      properties:
        name:
          type: string
          description: Post name
          pattern: '^[^\s]{1,40}$'
          example: my-Post
        description:
          type: string
          description: Post description
          pattern: '^[^\t\n]{0,255}$'
          example: A sample Post document
        status:
          type: string
          description: Post status
          example: active
...